js: Refactor for new release
